+++ src/prob/opf_iv.jl
@@ -1,0 +1,40 @@
+""
+function run_opf_iv(file, model_constructor, optimizer; kwargs...)
+    # return run_model(file, model_constructor, optimizer, post_opf_iv, solution_builder=solution_opf_iv!; kwargs...)
+    return run_model(file, model_constructor, optimizer, post_opf_iv; kwargs...)
+end
+
+""
+function post_opf_iv(pm::AbstractPowerModel)
+    variable_voltage(pm)
+    variable_branch_current(pm)
+
+    variable_gen(pm)
+    variable_dcline(pm)
+
+    objective_min_fuel_and_flow_cost(pm)
+
+    for i in ids(pm, :ref_buses)
+        constraint_theta_ref(pm, i)
+    end
+
+    for i in ids(pm, :bus)
+        constraint_current_balance(pm, i)
+    end
+
+    for i in ids(pm, :branch)
+        constraint_current_from(pm, i)
+        constraint_current_to(pm, i)
+
+        constraint_voltage_drop(pm, i)
+
+        constraint_voltage_angle_difference(pm, i)
+
+        constraint_thermal_limit_from(pm, i)
+        constraint_thermal_limit_to(pm, i)
+    end
+
+    for i in ids(pm, :dcline)
+        constraint_dcline(pm, i)
+    end
+end